interceptor/requirements.txt
# Assetto Corsa Race Interceptor Dependencies

# File system monitoring
watchdog>=3.0.0

# HTTP requests for REST API
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Socket.IO client for real-time updates
python-socketio[client]>=5.10.0
